lilas_api/lilas_api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
environs==11.0.0
fastapi==0.115.4
fastapi-jwt-auth==0.5.0
fastapi_cors==0.0.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
marshmallow==3.23.1
packaging==24.2
passlib==1.7.4
pycparser==2.22
pydantic==1.10.19
pydantic_core==2.23.4
PyJWT==1.7.1
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
#pip more lib
dependencies
redis
python-multipart
requests
python-dotenv
pytz
apscheduler

unidecode